# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API",
        "version": "0.1"
    },
    "paths": {
        "/api/signup": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a user throttling : 5",
                "operationId": "65b7ef1b1e076c033bf91055f53bd9a1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Log a user in throttling : 5",
                "operationId": "b5cee56757c3c79c5b5f115fdc33f460",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/signout": {
            "post": {
                "tags": [
                    "Logout"
                ],
                "summary": "Log a user out throttling : 5",
                "operationId": "b46f8611fc7d1123f2a3d573cee661c4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Authorization": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Register",
            "description": "Register"
        },
        {
            "name": "Login",
            "description": "Login"
        },
        {
            "name": "Logout",
            "description": "Logout"
        }
    ]
}